chore(kube-core): Move transformer to its own file.
Note: This should be done on other transformers in a future PR to keep this PR related to what it is fixing.
